Move ImageInfo to a new util.py file.
