+++ detect_supernova/pipeline.py
@@ -15,6 +15,7 @@ import subtraction
 import source_detection
 import truth_matching
 import truth_retrieval
+from util import ImageInfo
 
 
 def read_data_records(data_records_path):
@@ -429,7 +430,10 @@ def main():
                 "band": args.template_band,
             }
         else:
-            template_image_info = get_earliest_template_for_image(im)
+            science_image_path = Detection().INPUT_IMAGE_PATTERN.format(**science_id)
+            science_image = ImageInfo(science_image_path)
+
+            template_image_info = get_earliest_template_for_image(science_image)
             template_id = {
                 "pointing": template_image_info.pointing,
                 "sca": template_image_info.sca,
