+++ detect_supernova/subtraction.py
@@ -1,5 +1,4 @@
 import argparse
-from dataclasses import dataclass
 import gzip
 import io
 import os
@@ -14,6 +13,9 @@ import cupy as cp
 from sfft.SpaceSFFTCupyFlow import SpaceSFFT_CupyFlow
 from roman_imsim.utils import roman_utils
 from sfft.utils.SExSkySubtract import SEx_SkySubtract
+
+from util import ImageInfo
+
 
 INPUT_IMAGE_PATTERN = (
     "RomanTDS/images/simple_model/{band}/{pointing}/Roman_TDS_simple_model_{band}_{pointing}_{sca}.fits.gz"
@@ -132,24 +134,6 @@ def load_fits_to_cp(path, return_hdr=True, return_data=True, hdu_index=0, dtype=
             else None
         )
     return hdr, data
-
-
-@dataclass
-class ImageInfo:
-    data_id: dict
-    temp_dir: pathlib.Path
-
-    def __post_init__(self):
-        self.image_path = SIMS_DIR / pathlib.Path(
-            INPUT_IMAGE_PATTERN.format(**self.data_id)
-        )
-        self.cx = IMAGE_WIDTH // 2
-        self.cy = IMAGE_HEIGHT // 2
-
-        self.image_name = self.image_path.name
-        self.skysub_path = self.temp_dir / f"skysub_{self.image_name}"
-        self.detmask_path = self.temp_dir / f"detmask_{self.image_name}"
-        self.psf_path = self.temp_dir / f"psf_{self.image_name}"
 
 
 class Pipeline:
